+++ strand/iOS/Strand/Strand/DuffyCore/DFPeanutFeedObject.m
@@ -198,7 +198,7 @@ static NSArray *FeedObjectTypes;
 
 - (NSUInteger)hash
 {
-  return self.id;
+  return (NSUInteger)self.id;
 }
 
 - (NSArray *)actorAbbreviations
